requirements.txt
Django >= 1.11.0, < 1.12
djangorestframework >= 3.6.0, < 3.7
djangorestframework-jwt >= 1.11, < 2.0
django-cors-headers >= 2.1, < 3.0